requirements.txt
pytest
streamlit
sentence-transformers
faiss-cpu
numpy
